pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bpkit-cli"
version = "0.1.0"
description = "Transform business plans into executable MVP specifications for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "BP-Kit Contributors" }
]
keywords = ["business-plan", "specification", "ai-agents", "mvp", "constitution", "sequoia"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx[socks]>=0.25.0",
    "platformdirs>=3.0.0",
    "readchar>=4.0.0",
    "truststore>=0.10.4",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.0.0",
    "markdown-it-py>=3.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
bpkit = "bpkit_cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/bp-kit"
Repository = "https://github.com/yourusername/bp-kit"
Documentation = "https://github.com/yourusername/bp-kit#readme"
Issues = "https://github.com/yourusername/bp-kit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/bpkit_cli"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
